backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

torch
transformers>=4.40,<5

numpy>=1.26,<3
scipy>=1.11,<2

openai>=1.40.0
python-dotenv>=1.0.1
python-multipart>=0.0.9

Pillow>=10.0
pillow-heif>=0.16
